clean up to commit
